correctly implemented an ortho projection matrix
tried to implement it myself but was unable to make it work correctly

+++ src/renderer.cpp
@@ -1,7 +1,5 @@
 #include "renderer.h"
 
-#include <glm/matrix.hpp>
-
 #ifndef STB_IMAGE_IMPLEMENTATION
 #define STB_IMAGE_IMPLEMENTATION
 #include "stb_image.h"
@@ -39,8 +37,8 @@ bool loadTextureAtlas(char const *texturePath, GLContext *glContext,
     return true;
 }
 
-void draw_sprite(RenderData *renderData, SpriteID spriteID, glm::ivec2 pos,
-                 glm::ivec2 size) {
+void draw_sprite(RenderData *renderData, SpriteID spriteID, glm::vec2 pos,
+                 glm::vec2 size) {
 
     Sprite sp = get_sprite(spriteID);
